I have an erro when merging this to the order related branch

+++ Dtos/DishDto/AddDishDto.cs
@@ -6,7 +6,7 @@ namespace Arowolo_Delivery_Project.Dtos.DishDto
     {
         public string Name { get; set; } = "Rice";
         public string Description { get; set; } = "a good tasting food";
-        public int Price { get; set; } = 100;
+        public double Price { get; set; } = 100;
         public bool IsVegetarian { get; set; } = true;
         public Category Category { get; set; } = Category.Wok;
         public int Rating { get; set; } = 8;
